further fixes from rebases
