+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/types/dataset/mappers/DatasetUpdateInputMapper.java
@@ -2,12 +2,21 @@ package com.linkedin.datahub.graphql.types.dataset.mappers;
 
 import javax.annotation.Nonnull;
 
+import com.linkedin.common.GlobalTags;
+import com.linkedin.common.TagAssociation;
+import com.linkedin.common.TagAssociationArray;
+import com.linkedin.common.urn.TagUrn;
 import com.linkedin.datahub.graphql.generated.DatasetUpdateInput;
 import com.linkedin.datahub.graphql.types.common.mappers.InstitutionalMemoryUpdateMapper;
 import com.linkedin.datahub.graphql.types.common.mappers.OwnershipUpdateMapper;
 import com.linkedin.datahub.graphql.types.mappers.ModelMapper;
 import com.linkedin.dataset.Dataset;
 import com.linkedin.dataset.DatasetDeprecation;
+import com.linkedin.schema.EditableSchemaFieldInfo;
+import com.linkedin.schema.EditableSchemaFieldInfoArray;
+import com.linkedin.schema.EditableSchemaMetadata;
+
+import java.util.stream.Collectors;
 
 public class DatasetUpdateInputMapper implements ModelMapper<DatasetUpdateInput, Dataset> {
 
@@ -39,6 +48,40 @@ public class DatasetUpdateInputMapper implements ModelMapper<DatasetUpdateInput,
             result.setInstitutionalMemory(InstitutionalMemoryUpdateMapper.map(datasetUpdateInput.getInstitutionalMemory()));
         }
 
+        if (datasetUpdateInput.getGlobalTags() != null) {
+            final GlobalTags globalTags = new GlobalTags();
+            globalTags.setTags(new TagAssociationArray(datasetUpdateInput.getGlobalTags().getTags().stream().map(element -> mapTag(element)).collect(Collectors.toList())));
+            result.setGlobalTags(globalTags);
+        }
+
+        if (datasetUpdateInput.getEditableSchemaMetadata() != null) {
+            final EditableSchemaMetadata editableSchemaMetadata = new EditableSchemaMetadata();
+            editableSchemaMetadata.setEditableSchemaFieldInfo(
+                    new EditableSchemaFieldInfoArray(datasetUpdateInput.getEditableSchemaMetadata().getEditableSchemaFieldInfo().stream().map(element -> mapSchemaFieldInfo(element)).collect(Collectors.toList())));
+            result.setEditableSchemaMetadata(editableSchemaMetadata);
+
+        }
+
         return result;
+    }
+
+    private EditableSchemaFieldInfo mapSchemaFieldInfo(final com.linkedin.datahub.graphql.generated.EditableSchemaFieldInfoUpdate schemaFieldInfo) {
+        final EditableSchemaFieldInfo output = new EditableSchemaFieldInfo();
+        output.setDescription(schemaFieldInfo.getDescription());
+        output.setFieldPath(schemaFieldInfo.getFieldPath());
+
+        final GlobalTags globalTags = new GlobalTags();
+        globalTags.setTags(
+                new TagAssociationArray(schemaFieldInfo.getGlobalTags().getTags().stream().map(element -> mapTag(element)).collect(Collectors.toList()))
+        );
+        output.setGlobalTags(globalTags);
+
+        return output;
+    }
+
+    private TagAssociation mapTag(final com.linkedin.datahub.graphql.generated.TagAssociationUpdate tagAssociation) {
+        final TagAssociation output = new TagAssociation();
+        output.setTag(new TagUrn(tagAssociation.getTag().getUrn()));
+        return output;
     }
 }
