+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/types/dataset/mappers/SchemaFieldMapper.java
@@ -25,6 +25,9 @@ public class SchemaFieldMapper implements ModelMapper<com.linkedin.schema.Schema
         result.setNullable(input.isNullable());
         result.setNativeDataType(input.getNativeDataType());
         result.setType(mapSchemaFieldDataType(input.getType()));
+        if (input.hasGlobalTags()) {
+            result.setGlobalTags(GlobalTagsMapper.map(input.getGlobalTags()));
+        }
         return result;
     }
