+++ metadata-models/src/main/pegasus/com/linkedin/schema/SchemaField.pdl
@@ -1,6 +1,7 @@
 namespace com.linkedin.schema
 
 import com.linkedin.dataset.SchemaFieldPath
+import com.linkedin.common.GlobalTags
 
 /**
  * SchemaField to describe metadata related to dataset schema. Schema normalization rules: http://go/tms-schema
